requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
google-auth==2.38.0
google-genai==1.3.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==24.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
PySocks==1.7.1
requests==2.32.3
rsa==4.9
setuptools==80.9.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
Flask-Cors==4.0.0
pytest
